# artifacts/growweb/src/lib/api-client/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * GROWWEB Agency API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ContactInput {
  /** @minLength 1 */
  name: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  service?: string | null;
  /** @nullable */
  budget?: string | null;
  /** @minLength 1 */
  message: string;
}

export interface ContactReadUpdate {
  isRead?: boolean;
}

export interface ContactMessage {
  id: number;
  name: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  service?: string | null;
  /** @nullable */
  budget?: string | null;
  message: string;
  isRead: boolean;
  createdAt: string;
}

export interface PortfolioProject {
  id: number;
  title: string;
  category: string;
  description: string;
  imageUrl: string;
  technologies: string[];
  /** @nullable */
  liveUrl?: string | null;
  featured?: boolean;
  sortOrder?: number;
}

export interface PortfolioInput {
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  category: string;
  /** @minLength 1 */
  description: string;
  imageUrl: string;
  technologies: string[];
  /** @nullable */
  liveUrl?: string | null;
  featured?: boolean;
  sortOrder?: number;
}

export interface Testimonial {
  id: number;
  name: string;
  role: string;
  company: string;
  content: string;
  rating: number;
  /** @nullable */
  avatarUrl?: string | null;
}

export interface TestimonialInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  role: string;
  /** @minLength 1 */
  company: string;
  /** @minLength 1 */
  content: string;
  /**
     * @minimum 1
     * @maximum 5
     */
  rating: number;
  /** @nullable */
  avatarUrl?: string | null;
}

export interface AdminCredentials {
  username: string;
  password: string;
}

export interface AdminSession {
  token: string;
  username: string;
}

export interface DashboardStats {
  totalContacts: number;
  unreadContacts: number;
  totalProjects: number;
  totalTestimonials: number;
}

export interface SiteSettingsMap {[key: string]: string}

export type ListPortfolioParams = {
category?: string;
};