# lib/api-zod/src/generated/types/contactInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * GROWWEB Agency API
 * OpenAPI spec version: 0.1.0
 */

export interface ContactInput {
  /** @minLength 1 */
  name: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  service?: string | null;
  /** @nullable */
  budget?: string | null;
  /** @minLength 1 */
  message: string;
}