# lib/api-zod/src/generated/types/contactMessage.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * GROWWEB Agency API
 * OpenAPI spec version: 0.1.0
 */

export interface ContactMessage {
  id: number;
  name: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  service?: string | null;
  /** @nullable */
  budget?: string | null;
  message: string;
  isRead: boolean;
  createdAt: string;
}