# lib/api-zod/src/generated/types/portfolioInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * GROWWEB Agency API
 * OpenAPI spec version: 0.1.0
 */

export interface PortfolioInput {
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  category: string;
  /** @minLength 1 */
  description: string;
  imageUrl: string;
  technologies: string[];
  /** @nullable */
  liveUrl?: string | null;
  featured?: boolean;
  sortOrder?: number;
}