# lib/api-zod/src/generated/types/portfolioProject.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * GROWWEB Agency API
 * OpenAPI spec version: 0.1.0
 */

export interface PortfolioProject {
  id: number;
  title: string;
  category: string;
  description: string;
  imageUrl: string;
  technologies: string[];
  /** @nullable */
  liveUrl?: string | null;
  featured?: boolean;
  sortOrder?: number;
}