# lib/api-zod/src/generated/types/testimonial.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * GROWWEB Agency API
 * OpenAPI spec version: 0.1.0
 */

export interface Testimonial {
  id: number;
  name: string;
  role: string;
  company: string;
  content: string;
  rating: number;
  /** @nullable */
  avatarUrl?: string | null;
}